port to 64bit. Some gcc 4.4 warning fixes
git-svn-id: file:///home/notaz/opt/svn/PicoDrive@835 be3aeb3a-fb24-0410-a615-afba39da0efa

+++ cpu/cz80/cz80_op.c
@@ -917,7 +917,7 @@ OP_DJNZ:
 	OP(0x18):   // JR   n
 OP_JR:
 		adr = (INT8)READ_ARG();
-		PC += adr;
+		PC += (INT8)adr;
 		RET(12)
 
 	OP(0x20):   // JR   NZ,n
