+++ pico/sound/sound.c
@@ -291,7 +291,7 @@ PICO_INTERNAL void PsndClear(void)
     memset32((int *) PsndOut, 0, len); // assume PsndOut to be aligned
   else {
     short *out = PsndOut;
-    if ((int)out & 2) { *out++ = 0; len--; }
+    if ((long)out & 2) { *out++ = 0; len--; }
     memset32((int *) out, 0, len/2);
     if (len & 1) out[len-1] = 0;
   }
